refactor(auth): change AuthProviderEnum to avoid name collision with auth context
