+++ client/src/app/api/auth/[provider]/callback/route.ts
@@ -1,4 +1,4 @@
-import { authContainer, AuthProvider } from '@/features/auth'
+import { authContainer, AuthProviderEnum } from '@/features/auth/index.server'
 import { NextResponse } from 'next/server'
 
 const PARAM_CODE = 'code'
@@ -7,13 +7,13 @@ const PARAM_STATE = 'state'
 const HEADER_SET_COOKIE = 'Set-Cookie'
 
 export async function GET(request: Request, { params }: { params: Promise<{ provider: string }> }) {
-	const provider: AuthProvider = (await params).provider as AuthProvider
+	const provider: AuthProviderEnum = (await params).provider as AuthProviderEnum
 
 	const url = new URL(request.url)
 	const code = url.searchParams.get(PARAM_CODE)
 	const state = url.searchParams.get(PARAM_STATE)
 
-	const stateCheck = authContainer.service().verifyState(state || '')
+	const stateCheck = await authContainer.service().verifyState(state || '')
 	if (!stateCheck || !code) {
 		return NextResponse.redirect(new URL('/?error=no_code', request.url))
 	}
