+++ client/src/features/auth/auth-provider.enum.ts
@@ -1,4 +1,4 @@
-export enum AuthProvider {
+export enum AuthProviderEnum {
 	GOOGLE = 'google',
 	GITHUB = 'github',
 }
