+++ client/src/features/auth/index.client.ts
@@ -1,0 +1,1 @@
+export { AuthContext, useAuth, AuthProvider } from './auth.context'
