+++ client/src/features/auth/index.server.ts
@@ -1,3 +1,3 @@
 export { default as authContainer } from './auth.container'
 export type { default as IOAuthService } from './oauth.service.interface'
-export { AuthProvider } from './auth-provider.enum'
+export { AuthProviderEnum as AuthProviderEnum } from './auth-provider.enum'
