+++ client/src/features/auth/oauth.service.interface.ts
@@ -1,6 +1,6 @@
-import { AuthProvider } from './auth-provider.enum'
+import { AuthProviderEnum } from './auth-provider.enum'
 
 export default interface IOAuthService {
-	getAuthCodeUrlWithParams(provider: AuthProvider): Promise<string>
+	getAuthCodeUrlWithParams(provider: AuthProviderEnum): Promise<string>
 	verifyState(paramState: string): Promise<boolean>
 }
